build: improve test workflow (#11361)
