+++ .github/workflows/test/mongodb.ormconfig.json
@@ -4,23 +4,11 @@
     "disabledIfNotEnabledImplicitly": true,
     "name": "mongodb",
     "type": "mongodb",
-    "host": "mongodb",
+    "host": "localhost",
     "port": 27017,
     "database": "test",
     "logging": false,
     "useNewUrlParser": true,
     "useUnifiedTopology": true
-  },
-
-  {
-    "skip": true,
-    "name": "postgres",
-    "type": "postgres",
-    "host": "postgres",
-    "port": 5432,
-    "username": "postgres",
-    "password": "postgres",
-    "database": "postgres",
-    "logging": false
   }
 ]
