+++ .github/workflows/test/mysql-mariadb.ormconfig.json
@@ -3,7 +3,7 @@
     "skip": false,
     "name": "mysql",
     "type": "mysql",
-    "host": "mysql",
+    "host": "localhost",
     "port": 3306,
     "username": "root",
     "password": "admin",
@@ -14,23 +14,11 @@
     "skip": false,
     "name": "mariadb",
     "type": "mariadb",
-    "host": "mariadb",
-    "port": 3306,
+    "host": "localhost",
+    "port": 3307,
     "username": "root",
     "password": "admin",
     "database": "test",
-    "logging": false
-  },
-
-  {
-    "skip": true,
-    "name": "postgres",
-    "type": "postgres",
-    "host": "postgres",
-    "port": 5432,
-    "username": "postgres",
-    "password": "postgres",
-    "database": "postgres",
     "logging": false
   }
 ]
