+++ .github/workflows/test/oracle.ormconfig.json
@@ -5,24 +5,9 @@
     "type": "oracle",
     "host": "localhost",
     "port": 1521,
-    "serviceName": "XEPDB1",
+    "serviceName": "FREEPDB1",
     "username": "typeorm",
     "password": "oracle",
-    "logging": false,
-    "extra": {
-      "connectString": "localhost:1521/XEPDB1"
-    }
-  },
-
-  {
-    "skip": true,
-    "name": "postgres",
-    "type": "postgres",
-    "host": "postgres",
-    "port": 5432,
-    "username": "postgres",
-    "password": "postgres",
-    "database": "postgres",
     "logging": false
   }
 ]
