+++ .github/workflows/test/sqljs.ormconfig.json
@@ -4,17 +4,5 @@
     "name": "sqljs",
     "type": "sqljs",
     "logging": false
-  },
-
-  {
-    "skip": true,
-    "name": "postgres",
-    "type": "postgres",
-    "host": "postgres",
-    "port": 5432,
-    "username": "postgres",
-    "password": "postgres",
-    "database": "postgres",
-    "logging": false
   }
 ]
