+++ .mocharc.json
@@ -1,0 +1,8 @@
+{
+  "check-leaks": true,
+  "color": true,
+  "file": ["./build/compiled/test/utils/test-setup.js"],
+  "spec": ["./build/compiled/test"],
+  "timeout": "90000",
+  "recursive": true
+}
