--- test/functional/connection-options-reader/configs/yaml/test-yaml.yaml
@@ -1,3 +1,0 @@
-- type: "sqlite"
-  name: "file"
-  database: "test-yaml"
