+++ test/functional/connection/replication.ts
@@ -1,15 +1,16 @@
-import "reflect-metadata"
-import "../../utils/test-setup"
 import { expect } from "chai"
-import { Post } from "./entity/Post"
-import { Category } from "./entity/Category"
+import "reflect-metadata"
+
+import { QueryRunner } from "../../../src"
+import { DataSource } from "../../../src/data-source/DataSource"
+import "../../utils/test-setup"
 import {
     closeTestingConnections,
     createTestingConnections,
     getTypeOrmConfig,
 } from "../../utils/test-utils"
-import { DataSource } from "../../../src/data-source/DataSource"
-import { QueryRunner } from "../../../src"
+import { Category } from "./entity/Category"
+import { Post } from "./entity/Post"
 
 const expectCurrentApplicationName = async (
     queryRunner: QueryRunner,
@@ -22,18 +23,18 @@ const expectCurrentApplicationName = async (
 }
 
 describe("Connection replication", () => {
+    const ormConfigConnectionOptionsArray = getTypeOrmConfig()
+    const postgresOptions = ormConfigConnectionOptionsArray.find(
+        (options) => options.type == "postgres",
+    )
+    if (!postgresOptions) {
+        return
+    }
+
     describe("after connection is established successfully", function () {
         let connection: DataSource
-        beforeEach(async () => {
-            const ormConfigConnectionOptionsArray = getTypeOrmConfig()
-            const postgres = ormConfigConnectionOptionsArray.find(
-                (options) => options.type == "postgres",
-            )
-            if (!postgres)
-                throw new Error(
-                    "need a postgres connection in the test connection options to test replication",
-                )
 
+        beforeEach(async () => {
             connection = (
                 await createTestingConnections({
                     entities: [Post, Category],
@@ -42,14 +43,20 @@ describe("Connection replication", () => {
                     dropSchema: true,
                     driverSpecific: {
                         replication: {
-                            master: { ...postgres, applicationName: "master" },
-                            slaves: [{ ...postgres, applicationName: "slave" }],
+                            master: {
+                                ...postgresOptions,
+                                applicationName: "master",
+                            },
+                            slaves: [
+                                {
+                                    ...postgresOptions,
+                                    applicationName: "slave",
+                                },
+                            ],
                         },
                     },
                 })
             )[0]
-
-            if (!connection) return
 
             const post = new Post()
             post.title = "TypeORM Intro"
@@ -64,17 +71,11 @@ describe("Connection replication", () => {
 
         afterEach(() => closeTestingConnections([connection]))
 
-        it("connection.isConnected should be true", () => {
-            if (!connection || connection.driver.options.type !== "postgres") {
-                return
-            }
+        it("connection.isInitialized should be true", () => {
             connection.isInitialized.should.be.true
         })
 
         it("query runners should go to the master by default", async () => {
-            if (!connection || connection.driver.options.type !== "postgres") {
-                return
-            }
             const queryRunner = connection.createQueryRunner()
             expect(queryRunner.getReplicationMode()).to.equal("master")
 
@@ -83,9 +84,6 @@ describe("Connection replication", () => {
         })
 
         it("query runners can have their replication mode overridden", async () => {
-            if (!connection || connection.driver.options.type !== "postgres") {
-                return
-            }
             let queryRunner = connection.createQueryRunner("master")
             queryRunner.getReplicationMode().should.equal("master")
             await expectCurrentApplicationName(queryRunner, "master")
@@ -98,9 +96,6 @@ describe("Connection replication", () => {
         })
 
         it("read queries should go to the slaves by default", async () => {
-            if (!connection || connection.driver.options.type !== "postgres") {
-                return
-            }
             const result = await connection.manager
                 .createQueryBuilder(Post, "post")
                 .select("id")
@@ -113,9 +108,6 @@ describe("Connection replication", () => {
         })
 
         it("write queries should go to the master", async () => {
-            if (!connection || connection.driver.options.type !== "postgres") {
-                return
-            }
             const result = await connection.manager
                 .createQueryBuilder(Post, "post")
                 .insert()
@@ -132,16 +124,8 @@ describe("Connection replication", () => {
 
     describe("with custom replication default mode", function () {
         let connection: DataSource
-        beforeEach(async () => {
-            const ormConfigConnectionOptionsArray = getTypeOrmConfig()
-            const postgres = ormConfigConnectionOptionsArray.find(
-                (options) => options.type == "postgres",
-            )
-            if (!postgres)
-                throw new Error(
-                    "need a postgres connection in the test connection options to test replication",
-                )
 
+        beforeEach(async () => {
             connection = (
                 await createTestingConnections({
                     entities: [Post, Category],
@@ -151,14 +135,20 @@ describe("Connection replication", () => {
                     driverSpecific: {
                         replication: {
                             defaultMode: "master",
-                            master: { ...postgres, applicationName: "master" },
-                            slaves: [{ ...postgres, applicationName: "slave" }],
+                            master: {
+                                ...postgresOptions,
+                                applicationName: "master",
+                            },
+                            slaves: [
+                                {
+                                    ...postgresOptions,
+                                    applicationName: "slave",
+                                },
+                            ],
                         },
                     },
                 })
             )[0]
-
-            if (!connection) return
 
             const post = new Post()
             post.title = "TypeORM Intro"
@@ -174,9 +164,6 @@ describe("Connection replication", () => {
         afterEach(() => closeTestingConnections([connection]))
 
         it("query runners should go to the master by default", async () => {
-            if (!connection || connection.driver.options.type !== "postgres") {
-                return
-            }
             const queryRunner = connection.createQueryRunner()
             expect(queryRunner.getReplicationMode()).to.equal("master")
 
@@ -185,9 +172,6 @@ describe("Connection replication", () => {
         })
 
         it("query runners can have their replication mode overridden", async () => {
-            if (!connection || connection.driver.options.type !== "postgres") {
-                return
-            }
             let queryRunner = connection.createQueryRunner("master")
             queryRunner.getReplicationMode().should.equal("master")
             await expectCurrentApplicationName(queryRunner, "master")
@@ -200,9 +184,6 @@ describe("Connection replication", () => {
         })
 
         it("read queries should go to the master by default", async () => {
-            if (!connection || connection.driver.options.type !== "postgres") {
-                return
-            }
             const result = await connection.manager
                 .createQueryBuilder(Post, "post")
                 .select("id")
