+++ test/functional/find-options/array-contained-by-operator/array-contained-by-operator.test.ts
@@ -15,7 +15,6 @@ describe("find options > find operators > ArrayContainedBy", () => {
             (connections = await createTestingConnections({
                 __dirname,
                 enabledDrivers: ["postgres", "cockroachdb"],
-                // logging: true,
             })),
     )
     beforeEach(() => reloadTestingDatabases(connections))
