+++ test/functional/find-options/array-contains-operator/array-contains-operator.test.ts
@@ -1,12 +1,12 @@
-import "../../../utils/test-setup"
 import { DataSource, EntityManager } from "../../../../src"
+import { ArrayContains } from "../../../../src/find-options/operator/ArrayContains"
+import "../../../utils/test-setup"
 import {
     closeTestingConnections,
     createTestingConnections,
     reloadTestingDatabases,
 } from "../../../utils/test-utils"
 import { Post, PostStatus } from "./entity/Post"
-import { ArrayContains } from "../../../../src/find-options/operator/ArrayContains"
 
 describe("find options > find operators > ArrayContains", () => {
     let connections: DataSource[]
@@ -15,7 +15,6 @@ describe("find options > find operators > ArrayContains", () => {
             (connections = await createTestingConnections({
                 __dirname,
                 enabledDrivers: ["postgres", "cockroachdb"],
-                // logging: true,
             })),
     )
     beforeEach(() => reloadTestingDatabases(connections))
