+++ test/unit/connection-options-reader/configs/class-entities.ts
