+++ test/unit/connection-options-reader/configs/sqlite-memory.ts
