+++ test/unit/connection-options-reader/configs/test-path-config.ts
